PROD-2039 Extract functionality into hook
